move onconnectclickatom to wallet.ts
