+++ src/components/Swap/index.tsx
@@ -11,7 +11,8 @@ import useHasFocus from 'hooks/useHasFocus'
 import useOnSupportedNetwork from 'hooks/useOnSupportedNetwork'
 import { useAtom } from 'jotai'
 import { useEffect, useState } from 'react'
-import { displayTxHashAtom, onConnectWalletClickAtom } from 'state/swap'
+import { displayTxHashAtom } from 'state/swap'
+import { onConnectWalletClickAtom } from 'state/wallet'
 import { SwapTransactionInfo, Transaction, TransactionType, WrapTransactionInfo } from 'state/transactions'
 
 import Dialog from '../Dialog'
